Automatic Area Report of Celtic Invasive Plants database or selections (Area_Report_CIPdb)

Generate an automatic report of the taxa occurring in an administrative, protected or biogeographic area of the Celtic Invasive Plants database or a selection of this of the unique values of the categorical columns. The possible areas include: "Subprovince", "Country", "Constituent_Country_OR_Crown_Dependency", "Admin_units_II", "Admin_units_III", "National_Nature_Reserve_Name" and "Natura_2000_Name". This function retrieves the unique values for 22 categories obtained combining "Species_with_author" with the status and protected areas columns: "Species_AIS", "Species_AIS_&_EU_Concern", "Species_AIS_&_Listed", "Species_AIS_&_NOT_Listed", "Species_AIS_&_EU_Concern_&_Listed", "Species_AIS_&_EU_Concern_&_NOT_Listed","Species_AIS_&_Listed_in_Protected_Area", "Species_AIS_&_EU_Concern_&_Listed_in_Protected_Area","Species_AIS_&_Listed_in_National_Protected_Area", "Species_AIS_&_EU_Concern_&_National_Protected_Area", "Species_AIS_&_Listed_in_Natura_2000", "Species_AIS_&_EU_Concern_&_Natura_2000","Species_AIS_&_NOT_Listed_in_Protected_Area", "Species_AIS_&_EU_Concern_&_NOT_Listed_in_Protected_Area","Species_AIS_&_NOT_Listed_in_National_Protected_Area","Species_AIS_&_EU_Concern_&_NOT_LIsted_in_National_Protected_Area","Species_AIS_&_NOT_Listed_in_Natura_2000","Species_AIS_&_EU_Concern_&_NOT_Listed_in_Natura_2000","Species_NAIS_&_Native","Species_NAIS_&_Alien","Species_NAIS_&_Alien_&_Listed" and "Species_NAIS_&_Alien_&_NOT_Listed".

Area_Report_CIPdb(data, Scope, Values, output_name = deparse(substitute(data)))

Arguments

data

must be the table obtained from CIPdb() or another selection of this (Select_CIPdb, WDPA_PID_select_CIPdb).

Scope

must correspond with one of the possible area: Subprovince", "Country", "Constituent_Country_OR_Crown_Dependency", "Admin_units_II", "Admin_units_III", "National_Nature_Reserve_Name" and "Natura_2000_Name". The value in this argument must exactly correspond with the name of only one of these columns for this function to work.

Values

must be be either a character string or a vector of character strings.This must coincide with the potential values of the 'Scope' columns.

Value

This function returns a table object with 3 columns: Category, Number and Value. The Category columnn refers to: (1) area categories (Subprovince", "Country", "Constituent_Country_OR_Crown_Dependency", "Admin_units_II", "Admin_units_III", "National_Nature_Reserve_Name" and "Natura_2000_Name") and (2) taxa name and status and protected areas columns ("Species_AIS", "Species_AIS_&_EU_Concern", "Species_AIS_&_Listed", "Species_AIS_&_NOT_Listed", "Species_AIS_&_EU_Concern_&_Listed", "Species_AIS_&_EU_Concern_&_NOT_Listed","Species_AIS_&_Listed_in_Protected_Area", "Species_AIS_&_EU_Concern_&_Listed_in_Protected_Area","Species_AIS_&_Listed_in_National_Protected_Area", "Species_AIS_&_EU_Concern_&_National_Protected_Area", "Species_AIS_&_Listed_in_Natura_2000", "Species_AIS_&_EU_Concern_&_Natura_2000","Species_AIS_&_NOT_Listed_in_Protected_Area", "Species_AIS_&_EU_Concern_&_NOT_Listed_in_Protected_Area","Species_AIS_&_NOT_Listed_in_National_Protected_Area","Species_AIS_&_EU_Concern_&_NOT_LIsted_in_National_Protected_Area","Species_AIS_&_NOT_Listed_in_Natura_2000","Species_AIS_&_EU_Concern_&_NOT_Listed_in_Natura_2000","Species_NAIS_&_Native","Species_NAIS_&_Alien","Species_NAIS_&_Alien_&_Listed" and "Species_NAIS_&_Alien_&_NOT_Listed"). The Number column refers to the number of unique values of each category. The Values column refers to a list of the detected unique values in each category in which the different elements are separated by " & ". If the "Value" argument is fed with a vector containing more than one elements, this function will automatically create two dataframes (tables) named after the 'data' argument value and each of the vector element fed to the argument 'Value'.

Details

The optimal way of using this function is in combination with the Select_CIPdb() or WDPA_PID_select_CIPdb() functions.

Examples

#This provides an Area report for France
Data = CIPdb()
#> Loading the Celtic Invasive Plants database (CIPdb), this may take a few minutes...
#> CIPdb loaded successfully!
My_area_Report = Area_Report_CIPdb(Data, "Country", c("France"))
#This provides an Area report for France  and Spain
Data = CIPdb()
#> Loading the Celtic Invasive Plants database (CIPdb), this may take a few minutes...
#> CIPdb loaded successfully!
My_countries = c("Spain / España", "France")
#This will generate  a "Spain / España" dataframe and a "France" dataframe
My_area_Report = Area_Report_CIPdb(Data, "Country", My_countries)
#This provides an Area report for the Cantabrian Atlantic subprovince
My_area_Report = Area_Report_CIPdb(Data, "Subprovince", "Cantabrian Atlantic")
#This provides an Area report for the Cantabrian Atlantic subprovince in France
Country_subprov = c("Cantabrian Atlantic", "France")
My_selection <- Select_CIPdb(Data, query = Country_subprov)
My_area_Report = Area_Report_CIPdb(My_selection, "Subprovince", "Cantabrian Atlantic")
#This provides an Area report for Wales and Scotland
W_S = c("Wales / Cymru", "Scotland / Alba")
My_area_Report = Area_Report_CIPdb(Data, "Constituent_Country_OR_Crown_Dependency", W_S)
#This provides an Area report of Asturias, York, Devon and Dordogne
admin_III = c("Asturias / Asturies", "York", "Dordogne", "Devon" )
My_area_Report = Area_Report_CIPdb(Data, "Admin_units_III", admin_III )