Native & Invasive Alien Species (NAIS) within the Celtic Invasive Plants Checklist verification (CIP_NAIS_Ver)

Native & Invasive Alien Species (NAIS) within the Celtic Invasive Plants Checklist verification (CIP_NAIS_Ver)

CIP_NAIS_Ver(
  url =
    "https://zenodo.org/records/18630660/files/CIP_DB_APRIL_2025_NAIS_VERIFICATION.csv?download=1"
)

Value

A table detailing the Native & Invasive Alien Species (NAIS) found within within the Celtic Invasive Plants Checklist and the references used to verify their native status by country.

References

Börgesen, F. (1908). Gardening and Tree-planting. In: Botany of the Faeröes, based upon Danish investigations Part III. Nordisk Forlag. Copenhagen (Denmark), pp. 1027-1043. https://doi.org/10.5962/bhl.title.8101 Castroviejo, S., Laínz, M., López González, G., Montserrat, P., Muñoz Garmendia, F., Paiva, J. & Villar, L. (1986-2020) Flora iberica. Plantas vasculares de la Peninsula Ibérica e Islas Baleares. Madrid, Real Jardín Botánico de Madrid, Consejo Superior de Investigaciones Científicas (CSIC). Domínguez Lozano, F. (2000). Atlas y Libro Rojo de la Flora Vascular Amenazada de España. Conservación vegetal 6. https://www.miteco.gob.es/content/dam/miteco/es/biodiversidad/temas/inventarios-nacionales/lista_roja_2000_tcm30-99751.pdf Fosaa, A. M. (2001). A Review of the Plant Communities of the Faroe Islands. Fróðskaparrit - Faroese Scientific Journal, 48, 41-54. https://doi.org/10.18602/fsj.v48i.756 Henniges, M. C., Powell, R. F., Mian, S., Stace, C. A., Walker, K. J., Gornall, R. J., ... & Leitch, I. J. (2022). A taxonomic, genetic and ecological data resource for the vascular plants of Britain and Ireland. Scientific Data, 9(1), 1. https://doi.org/10.1038/s41597-021-01104-5 Ostenfeld, C.H. (1901). Flora of the Faeröes: Phanerogamae and Pteridophyta. In: Botany of the Faeröes, based upon Danish investigations Part I. Nordisk Forlag. Copenhagen (Denmark), pp 41-100. https://doi.org/10.5962/bhl.title.8101 Ostenfeld, C.H. (1908). Additions and corrections of the List of Phanerogamae and Pteridophyta of Faeröes In: Botany of the Faeröes, based upon Danish investigations Part III. Nordisk Forlag. Copenhagen (Denmark), pp. 835-864. https://www.biodiversitylibrary.org/page/8496529 Royal Botanic Gardens Kew (2025) Plants of the World Online (POWO). Facilitated by the Royal Botanic Gardens, Kew. 2023. http://www.plantsoftheworldonline.org/ [Accessed: 1 April 2025]. Ramos-Gutiérrez, I., Lima, H., Pajarón, S., Romero-Zarco, C., Sáez, L., Pataro, L., Molina-Venegas, R., Rodríguez, M.A. & Moreno-Saiz, J.C. (2021) Atlas of the vascular flora of the Iberian Peninsula biodiversity hotspot (AFLIBER). Global Ecology and Biogeography. 30, 1951–1957. https://doi.org/10.1111/geb.13363 Tela Botanica (2022) Tela Botanica–Les bases de données botaniques. 2022. https://www.tela-botanica.org/ressources/donnees/telechargements/#donnes-observation [Accessed: 14 September 2022].

Examples

NAIS_ver <- CIP_NAIS_Ver()
#> Loading the Native & Invasive Alien Species (NAIS) within the Celtic Invasive Plants Checklist verification (CIP_NAIS_Ver), this may take a few minutes...
#> CIP_NAIS_Ver loaded successfully!
head(NAIS_ver)
#>   Taxa_ID                                                            Taxa_URL
#> 1  749646 https://powo.science.kew.org/taxon/urn:lsid:ipni.org:names:529047-1
#> 2  557362 https://powo.science.kew.org/taxon/urn:lsid:ipni.org:names:328527-2
#> 3  759344  https://powo.science.kew.org/taxon/urn:lsid:ipni.org:names:86508-1
#> 4  820195 https://powo.science.kew.org/taxon/urn:lsid:ipni.org:names:140949-1
#> 5  603933 https://powo.science.kew.org/taxon/urn:lsid:ipni.org:names:319145-2
#> 6  376033 https://powo.science.kew.org/taxon/urn:lsid:ipni.org:names:305797-1
#>          Phylum   Subphylum        Order         Family     Genus
#> 1 Spermatophyta Angiosperms  Asparagales Amaryllidaceae    Allium
#> 2 Spermatophyta Angiosperms Ranunculales  Berberidaceae  Berberis
#> 3 Spermatophyta Angiosperms  Alismatales        Araceae     Calla
#> 4 Spermatophyta Angiosperms    Asterales  Campanulaceae Campanula
#> 5 Spermatophyta Angiosperms Ranunculales  Ranunculaceae  Clematis
#> 6 Spermatophyta Angiosperms       Poales     Cyperaceae   Cyperus
#>          Species_with_Author                    Taxa Author
#> 1       Allium triquetrum L.       Allium triquetrum     L.
#> 2       Berberis vulgaris L.       Berberis vulgaris     L.
#> 3         Calla palustris L.         Calla palustris     L.
#> 4 Campanula rapunculoides L. Campanula rapunculoides     L.
#> 5        Clematis vitalba L.        Clematis vitalba     L.
#> 6        Cyperus rotundus L.        Cyperus rotundus     L.
#>   Celtic_Fringe_Taxa_Category Celtic_Fringe_Origin
#> 1                        NAIS               Native
#> 2                        NAIS               Native
#> 3                        NAIS               Native
#> 4                        NAIS               Native
#> 5                        NAIS               Native
#> 6                        NAIS               Native
#>                                                                                                        Status__Spain
#> 1            Castroviejo (1986-2020) [Native] & Ramos-Gutiérrez et al (2021) [Native] & POWO (2025) [Native]: Native
#> 2            Castroviejo (1986-2020) [Native] & Ramos-Gutiérrez et al (2021) [Native] & POWO (2025) [Native]: Native
#> 3 Castroviejo (1986-2020) [by absence] & Ramos-Gutiérrez et al (2021) [by absence] & POWO (2025) [by absence]: Alien
#> 4            Castroviejo (1986-2020) [Native] & Ramos-Gutiérrez et al (2021) [Native] & POWO (2025) [Native]: Native
#> 5                                                    Castroviejo (1986-2020) [Native] & POWO (2025) [Native]: Native
#> 6            Castroviejo (1986-2020) [Native] & Ramos-Gutiérrez et al (2021) [Native] & POWO (2025) [Native]: Native
#>                                                                                                                                          Status__Denmark
#> 1                                                      Ostenfeld (1901, 1908) [by absence] & Fosaa (2001) [by absence] & POWO (2025) [by absence]: Alien
#> 2                                                      Ostenfeld (1901, 1908) [by absence] & Fosaa (2001) [by absence] & POWO (2025) [by absence]: Alien
#> 3                                                      Ostenfeld (1901, 1908) [by absence] & Fosaa (2001) [by absence] & POWO (2025) [by absence]: Alien
#> 4 Ostenfeld (1901, 1908) [by absence] & Börgesen (1908) [as genus “Campanulla”, Planted] & Fosaa (2001) [by absence] & & POWO (2025) [by absence]: Alien
#> 5                                                    Ostenfeld (1901, 1908) [by absence] & Fosaa (2001) [by absence]  & POWO (2025) [by absence]:  Alien
#> 6                                                      Ostenfeld (1901, 1908) [by absence] & Fosaa (2001) [by absence] & POWO (2025) [by absence]: Alien
#>                                                                                                     Status__UK
#> 1 Henniges et al. (2022)  [“Neophyte naturalized” & “Neophyte”] & POWO (2025)  [“Introducted”]: Alien & Listed
#> 2                       Henniges et al. (2022)  [“Neophyte naturalized”] & POWO (2025)  [“Introducted”]: Alien
#> 3 Henniges et al. (2022)  [“Neophyte naturalized” & “Neophyte”] & POWO (2025)  [“Introducted”]: Alien & Listed
#> 4          Henniges et al. (2022)  [“Neophyte naturalized” & “Neophyte”] & POWO (2025)  [“Introducted”]: Alien
#> 5                                            Henniges et al. (2022)  [“Native”] & POWO (2025) [Native]: Native
#> 6                                                                 POWO (2025)  [“Introducted”]: Alien & Listed
#>                                                                                                Status__Ireland
#> 1 Henniges et al. (2022)  [“Neophyte naturalized” & “Neophyte”] & POWO (2025)  [“Introducted”]: Alien & Listed
#> 2              Henniges et al. (2022)  [“Neophyte naturalized”] & POWO (2025)  [“Introducted”]: Alien & Listed
#> 3         Henniges et al. (2022)  [“Neophyte naturalized” & “Neophyte”] &  POWO (2025)  [“Introducted”]: Alien
#> 4 Henniges et al. (2022)  [“Neophyte naturalized” & “Neophyte”] & POWO (2025)  [“Introducted”]: Alien & Listed
#> 5                            Henniges et al. (2022)  [“Native”] & POWO (2025)  [“Introducted”]: Alien & Listed
#> 6                                                                              POWO (2025) [by absence]: Alien
#>                                                                                  Status__France
#> 1 Tela Botanica (2022) [“N-Natif (=indigène)” & “Non introduit”] & POWO (2025) [Native]: Native
#> 2 Tela Botanica (2022) [“N-Natif (=indigène)” & “Non introduit”] & POWO (2025) [Native]: Native
#> 3 Tela Botanica (2022) [“N-Natif (=indigène)” & “Non introduit”] & POWO (2025) [Native]: Native
#> 4 Tela Botanica (2022) [“N-Natif (=indigène)” & “Non introduit”] & POWO (2025) [Native]: Native
#> 5 Tela Botanica (2022) [“N-Natif (=indigène)” & “Non introduit”] & POWO (2025) [Native]: Native
#> 6                          Tela Botanica (2022) [“A-Non natif”] & POWO (2025) [“Native”]: Alien
#>                                                                                                         Status__Portugal
#> 1                Castroviejo (1986-2020) [Native] & Ramos-Gutiérrez et al (2021) [Native] & POWO (2025) [Native]: Native
#> 2 Castroviejo (1986-2020) [by absence] & Ramos-Gutiérrez et al (2021) [by absence] & POWO (2025)  [“Introducted”]: Alien
#> 3     Castroviejo (1986-2020) [by absence] & Ramos-Gutiérrez et al (2021) [by absence] & POWO (2025) [by absence]: Alien
#> 4     Castroviejo (1986-2020) [by absence] & Ramos-Gutiérrez et al (2021) [by absence] & POWO (2025) [by absence]: Alien
#> 5                                                        Castroviejo (1986-2020) [Native] & POWO (2025) [Native]: Native
#> 6       Castroviejo (1986-2020) [Native] & Ramos-Gutiérrez et al (2021) [Native] & POWO (2025) [Native]: Native & Listed