Celtic Invasive Plants database (CIPdb)

Celtic Invasive Plants database (CIPdb)

CIPdb(
  url = "https://zenodo.org/records/18630660/files/CIP_DB_APRIL_2025.csv?download=1"
)

Value

This function automatically downloads the Celtic Invasive Plants database as table object called CIPdb.

References

Static version: González-Toral, C., Madrazo-Frías, L., Estrada Fernández, A., López-Alonso, R., Sanna, M., Cuesta, C., Cires, E. & Viruel, J. (202) Celtic Invasive Plants database. Version December 2025. Zenodo.org https://doi.org/10.5281/zenodo.17871899 R package version December 2025: González-Toral, C., Madrazo-Frías, L., Estrada Fernández, A., López-Alonso, R., Sanna, M., Cuesta, C., Cires, E. & Viruel, J. (202) Celtic Invasive Plants database. Version December 2025. Zenodo.org https://doi.org/10.5281/zenodo.17871899

Examples

Data = CIPdb()
#> Loading the Celtic Invasive Plants database (CIPdb), this may take a few minutes...
#> CIPdb loaded successfully!
head(Data)
#>   Taxa_ID                                                             Taxa_URL
#> 1  505938 https://powo.science.kew.org/taxon/urn:lsid:ipni.org:names:1044174-2
#> 2  505938 https://powo.science.kew.org/taxon/urn:lsid:ipni.org:names:1044174-2
#> 3  505938 https://powo.science.kew.org/taxon/urn:lsid:ipni.org:names:1044174-2
#> 4  505938 https://powo.science.kew.org/taxon/urn:lsid:ipni.org:names:1044174-2
#> 5  505938 https://powo.science.kew.org/taxon/urn:lsid:ipni.org:names:1044174-2
#> 6  505938 https://powo.science.kew.org/taxon/urn:lsid:ipni.org:names:1044174-2
#>          Phylum   Subphylum    Order    Family    Genus
#> 1 Spermatophyta Angiosperms Malvales Malvaceae Abutilon
#> 2 Spermatophyta Angiosperms Malvales Malvaceae Abutilon
#> 3 Spermatophyta Angiosperms Malvales Malvaceae Abutilon
#> 4 Spermatophyta Angiosperms Malvales Malvaceae Abutilon
#> 5 Spermatophyta Angiosperms Malvales Malvaceae Abutilon
#> 6 Spermatophyta Angiosperms Malvales Malvaceae Abutilon
#>           Species_with_Author Officially_listed EU_Concern Local_Origin
#> 1 Abutilon theophrasti Medik.                NO         NO        Alien
#> 2 Abutilon theophrasti Medik.                NO         NO        Alien
#> 3 Abutilon theophrasti Medik.                NO         NO        Alien
#> 4 Abutilon theophrasti Medik.                NO         NO        Alien
#> 5 Abutilon theophrasti Medik.                NO         NO        Alien
#> 6 Abutilon theophrasti Medik.                NO         NO        Alien
#>   Celtic_Fringe_Origin Celtic_Fringe_Taxa_Category                 Taxa
#> 1                Alien                         AIS Abutilon theophrasti
#> 2                Alien                         AIS Abutilon theophrasti
#> 3                Alien                         AIS Abutilon theophrasti
#> 4                Alien                         AIS Abutilon theophrasti
#> 5                Alien                         AIS Abutilon theophrasti
#> 6                Alien                         AIS Abutilon theophrasti
#>   UTM_grid         Subprovince Latitude  Longitude Country
#> 1  30TXN59 Cantabrian Atlantic 43.24788 -1.1522727  France
#> 2  30TXQ98 Cantabrian Atlantic 44.94804 -0.5914788  France
#> 3  30TYQ17 Cantabrian Atlantic 44.85249 -0.3423593  France
#> 4  30TXN67 Cantabrian Atlantic 43.06584 -1.0349382  France
#> 5  30TYN49 Cantabrian Atlantic 43.22462 -0.0448116  France
#> 6  30TYQ16 Cantabrian Atlantic 44.76256 -0.3464883  France
#>   Constituent_Country_OR_Crown_Dependency     Admin_units_II
#> 1                                    <NA> Nouvelle-Aquitaine
#> 2                                    <NA> Nouvelle-Aquitaine
#> 3                                    <NA> Nouvelle-Aquitaine
#> 4                                    <NA> Nouvelle-Aquitaine
#> 5                                    <NA>          Occitanie
#> 6                                    <NA> Nouvelle-Aquitaine
#>        Admin_units_III Presence_Protected_Area Presence_Natura_2000
#> 1 Pyrénées-Atlantiques                       1                    1
#> 2              Gironde                       1                    1
#> 3              Gironde                       1                    1
#> 4 Pyrénées-Atlantiques                       1                    1
#> 5      Hautes-Pyrénées                       0                    0
#> 6              Gironde                       1                    1
#>   Presence_National_Nature_Reserve National_Nature_Reserve_Name
#> 1                                0                         <NA>
#> 2                                1                        Médoc
#> 3                                0                         <NA>
#> 4                                0                         <NA>
#> 5                                0                         <NA>
#> 6                                0                         <NA>
#>                                                 Natura_2000_Name
#> 1                                       La Bidouze (cours d'eau)
#> 2                               La Garonne en Nouvelle-Aquitaine
#> 3                                Réseau hydrographique du Gestas
#> 4 Haute Soule : forêt d'Iraty, Orgambidexka et Pic des Escaliers
#> 5                                                           <NA>
#> 6                            Réseau hydrographique de la Pimpine
#>                  National_Designation             Designation_in_English
#> 1                         Natura 2000                        Natura 2000
#> 2 Parc naturel régional & Natura 2000 Regional Nature Park & Natura 2000
#> 3                         Natura 2000                        Natura 2000
#> 4                         Natura 2000                        Natura 2000
#> 5                                <NA>                               <NA>
#> 6                         Natura 2000                        Natura 2000
#>      Designation_Type  UNEP_WCMC_Cat     IUCN_Cat
#> 1            Regional Not Applicable Not Reported
#> 2 Regional & National Not Applicable            V
#> 3            Regional Not Applicable Not Reported
#> 4            Regional Not Applicable Not Reported
#> 5                <NA>           <NA>         <NA>
#> 6            Regional Not Applicable Not Reported
#>                          Governance_Type
#> 1 Federal or national ministry or agency
#> 2 Federal or national ministry or agency
#> 3 Federal or national ministry or agency
#> 4 Federal or national ministry or agency
#> 5                                   <NA>
#> 6 Federal or national ministry or agency
#>                                                                                                                                                                                                                                                                                                                                                                                                                   Management_Authority
#> 1                                                                                                                                                                                                                                                            See European Environment Agency Geospatial Data Catalogue for more information - https://sdi.eea.europa.eu/catalogue/srv/api/records/be2142b0-7dc4-42e3-afac-f7cc5f1a9ac6
#> 2 See European Environment Agency Geospatial Data Catalogue for more information - https://sdi.eea.europa.eu/catalogue/srv/api/records/be2142b0-7dc4-42e3-afac-f7cc5f1a9ac6 & Full information available from the EEA website - French Ministry in charge of the environment. In march 2012, name, acronym and address as follows : Ministère de l'Écologie, du Développement durable, des Transports et du Logement (MEDDTL) 92055 La
#> 3                                                                                                                                                                                                                                                            See European Environment Agency Geospatial Data Catalogue for more information - https://sdi.eea.europa.eu/catalogue/srv/api/records/be2142b0-7dc4-42e3-afac-f7cc5f1a9ac6
#> 4                                                                                                                                                                                                                                                            See European Environment Agency Geospatial Data Catalogue for more information - https://sdi.eea.europa.eu/catalogue/srv/api/records/be2142b0-7dc4-42e3-afac-f7cc5f1a9ac6
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                 <NA>
#> 6                                                                                                                                                                                                                                                            See European Environment Agency Geospatial Data Catalogue for more information - https://sdi.eea.europa.eu/catalogue/srv/api/records/be2142b0-7dc4-42e3-afac-f7cc5f1a9ac6
#>   Special_Areas_of_Conservation_.Habitats_Directive.
#> 1                                                YES
#> 2                                                YES
#> 3                                                YES
#> 4                                                 NO
#> 5                                               <NA>
#> 6                                                YES
#>   Special_Protection_Area_.Birds_Directive. Reserve_Ecosystem_Type
#> 1                                        NO            Terrestrial
#> 2                                        NO  Coastal & Terrestrial
#> 3                                        NO            Terrestrial
#> 4                                       YES            Terrestrial
#> 5                                      <NA>                   <NA>
#> 6                                        NO            Terrestrial
#>   Site_of_Community_Importance_.Habitats_Directive.
#> 1                                              <NA>
#> 2                                              <NA>
#> 3                                              <NA>
#> 4                                              <NA>
#> 5                                               YES
#> 6                                              <NA>
#>   Regional_Protected_Landscape_WDPA_PID Local_Nature_Reserve_WDPA_PID
#> 1                                   NaN                           NaN
#> 2                                   NaN                           NaN
#> 3                                   NaN                           NaN
#> 4                                   NaN                           NaN
#> 5                                   NaN                           NaN
#> 6                                   NaN                           NaN
#>   Marine_Protected_Area_.OSPAR._WDPA_PID Nature_Park_WDPA_PID
#> 1                                   <NA>                  NaN
#> 2                                   <NA>                  NaN
#> 3                                   <NA>                  NaN
#> 4                                   <NA>                  NaN
#> 5                                   <NA>                  NaN
#> 6                                   <NA>                  NaN
#>   Site_of_Community_Importance_.Habitats_Directive._WDPA_PID
#> 1                                                        NaN
#> 2                                                        NaN
#> 3                                                        NaN
#> 4                                                        NaN
#> 5                                                        NaN
#> 6                                                        NaN
#>   National_Park_WDPA_PID Protected_Landscape_WDPA_PID Natural_Park_WDPA_PID
#> 1                    NaN                          NaN                   NaN
#> 2                    NaN                          NaN                   NaN
#> 3                    NaN                          NaN                   NaN
#> 4                    NaN                          NaN                   NaN
#> 5                    NaN                          NaN                   NaN
#> 6                    NaN                          NaN                   NaN
#>   Natura_2000_WDPA_PID
#> 1                 <NA>
#> 2                 <NA>
#> 3                 <NA>
#> 4                 <NA>
#> 5                 <NA>
#> 6                 <NA>
#>   Special_Areas_of_Conservation_.Habitats_Directive._WDPA_PID
#> 1                                                   555788709
#> 2                                                   555788021
#> 3                                                   555787828
#> 4                                                        <NA>
#> 5                                                        <NA>
#> 6                                                   555526352
#>   Special_Protection_Area_.Birds_Directive._WDPA_PID Natural_Monument_WDPA_PID
#> 1                                               <NA>                       NaN
#> 2                                               <NA>                       NaN
#> 3                                               <NA>                       NaN
#> 4                                          555539428                       NaN
#> 5                                               <NA>                       NaN
#> 6                                               <NA>                       NaN
#>   Protected_Wetland_WDPA_PID Nature_Reserve_WDPA_PID
#> 1                        NaN                     NaN
#> 2                        NaN                     NaN
#> 3                        NaN                     NaN
#> 4                        NaN                     NaN
#> 5                        NaN                     NaN
#> 6                        NaN                     NaN
#>   Regional_Nature_Park_WDPA_PID Regional_Park_WDPA_PID
#> 1                           NaN                    NaN
#> 2                     555597294                    NaN
#> 3                           NaN                    NaN
#> 4                           NaN                    NaN
#> 5                           NaN                    NaN
#> 6                           NaN                    NaN
#>   National_Park_._Buffer_zone.Area_of_adhesion_WDPA_PID
#> 1                                                   NaN
#> 2                                                   NaN
#> 3                                                   NaN
#> 4                                                   NaN
#> 5                                                   NaN
#> 6                                                   NaN
#>   Biotope_Protection_Order_WDPA_PID Nature_enclave_WDPA_PID
#> 1                              <NA>                     NaN
#> 2                              <NA>                     NaN
#> 3                              <NA>                     NaN
#> 4                              <NA>                     NaN
#> 5                              <NA>                     NaN
#> 6                              <NA>                     NaN
#>   Land_acquired_by_a_regional_conservatory_of_natural_areas_WDPA_PID
#> 1                                                                NaN
#> 2                                                                NaN
#> 3                                                                NaN
#> 4                                                                NaN
#> 5                                                                NaN
#> 6                                                                NaN
#>   Regional_Nature_Reserve_WDPA_PID UNESCO.MAB_Biosphere_Reserve_WDPA_PID
#> 1                              NaN                                   NaN
#> 2                              NaN                                   NaN
#> 3                              NaN                                   NaN
#> 4                              NaN                                   NaN
#> 5                              NaN                                   NaN
#> 6                              NaN                                   NaN
#>   Natural_Area_.recreational._WDPA_PID Biosphere_Reserve_WDPA_PID
#> 1                                  NaN                        NaN
#> 2                                  NaN                        NaN
#> 3                                  NaN                        NaN
#> 4                                  NaN                        NaN
#> 5                                  NaN                        NaN
#> 6                                  NaN                        NaN
#>   Nature_Reserve_.Parcial._WDPA_PID Natural_Area_of_Special_Interest_WDPA_PID
#> 1                               NaN                                       NaN
#> 2                               NaN                                       NaN
#> 3                               NaN                                       NaN
#> 4                               NaN                                       NaN
#> 5                               NaN                                       NaN
#> 6                               NaN                                       NaN
#>   Land_acquired_by_Conservatoire_du_Littoral_.national_seaside_and_lakeside_conservancy._WDPA_PID
#> 1                                                                                             NaN
#> 2                                                                                             NaN
#> 3                                                                                             NaN
#> 4                                                                                             NaN
#> 5                                                                                             NaN
#> 6                                                                                             NaN
#>   Ramsar_Site._Wetland_of_International_Importance_WDPA_PID
#> 1                                                       NaN
#> 2                                                       NaN
#> 3                                                       NaN
#> 4                                                       NaN
#> 5                                                       NaN
#> 6                                                       NaN
#>   National_Nature_Reserve_WDPA_PID Site_of_national_interest_WDPA_PID
#> 1                              NaN                                NaN
#> 2                              NaN                                NaN
#> 3                              NaN                                NaN
#> 4                              NaN                                NaN
#> 5                              NaN                                NaN
#> 6                              NaN                                NaN
#>   Marine_Nature_Park_WDPA_PID
#> 1                         NaN
#> 2                         NaN
#> 3                         NaN
#> 4                         NaN
#> 5                         NaN
#> 6                         NaN
#>   Protected_perimeter_around_a_national_nature_reserve_WDPA_PID
#> 1                                                           NaN
#> 2                                                           NaN
#> 3                                                           NaN
#> 4                                                           NaN
#> 5                                                           NaN
#> 6                                                           NaN
#>   Forest_Managed_Biological_Reserve_WDPA_PID
#> 1                                        NaN
#> 2                                        NaN
#> 3                                        NaN
#> 4                                        NaN
#> 5                                        NaN
#> 6                                        NaN
#>   Forest_Integral_Biological_Reserve_WDPA_PID Geotope_Protection_Order_WDPA_PID
#> 1                                         NaN                              <NA>
#> 2                                         NaN                              <NA>
#> 3                                         NaN                              <NA>
#> 4                                         NaN                              <NA>
#> 5                                         NaN                              <NA>
#> 6                                         NaN                              <NA>
#>   Natural_habitats_protection_Order_WDPA_PID
#> 1                                        NaN
#> 2                                        NaN
#> 3                                        NaN
#> 4                                        NaN
#> 5                                        NaN
#> 6                                        NaN
#>   National_Hunting_and_Wildlife_Reserve_WDPA_PID Natural_Heritage_Area_WDPA_PID
#> 1                                            NaN                            NaN
#> 2                                            NaN                            NaN
#> 3                                            NaN                            NaN
#> 4                                            NaN                            NaN
#> 5                                            NaN                            NaN
#> 6                                            NaN                            NaN